Codigo/BackEnd/requirements.txt
# requirements.txt 

# FastAPI e servidor ASGI
fastapi==0.115.13   # Versão recomendada para compatibilidade atual
uvicorn==0.34.3    # Versão recomendada para o servidor ASGI

# Groq para LLM
groq>=0.8.0

# Processamento de PDF
PyPDF2==3.0.1

# Variáveis de Ambiente
python-dotenv==1.0.0

# Para requisições HTTP (geralmente usado implicitamente por libs, mas bom ter)
requests==2.31.0

# Para deploy em ambiente de produção (servidor WSGI)
gunicorn==21.2.0

# Para autenticação, se você usar OAuth2PasswordRequestForm
python-multipart==0.0.9 # Necessário para o FastAPI processar form-data
